+++ src/agent/netlink/src/lib.rs
@@ -9,7 +9,7 @@
 #![allow(unused_parens)]
 #![allow(unused_unsafe)]
 
-mod neigh;
+mod agent_handler;
 
 extern crate libc;
 extern crate nix;
@@ -26,8 +26,6 @@ extern crate slog_scope;
 extern crate scan_fmt;
 
 use nix::errno::Errno;
-use protobuf::RepeatedField;
-use protocols::types::{IPAddress, IPFamily, Interface, Route};
 use std::clone::Clone;
 use std::default::Default;
 use std::fmt;
@@ -1098,6 +1096,44 @@ pub const IFA_F_STABLE_PRIVACY: __u32 = 0x800;
 
 #[repr(C)]
 #[derive(Copy)]
+pub struct ndmsg {
+    ndm_family: __u8,
+    ndm_pad1: __u8,
+    ndm_pad: __u16,
+    ndm_ifindex: __s32,
+    ndm_state: __u16,
+    ndm_flags: __u8,
+    ndm_type: __u8,
+}
+
+pub const NDA_UNSPEC: __u16 = 0;
+pub const NDA_DST: __u16 = 1;
+pub const NDA_LLADDR: __u16 = 2;
+pub const NDA_CACHEINFO: __u16 = 3;
+pub const NDA_PROBES: __u16 = 4;
+pub const NDA_VLAN: __u16 = 5;
+pub const NDA_PORT: __u16 = 6;
+pub const NDA_VNI: __u16 = 7;
+pub const NDA_IFINDEX: __u16 = 8;
+pub const NDA_MASTER: __u16 = 9;
+pub const NDA_LINK_NETNSID: __u16 = 10;
+pub const NDA_SRC_VNI: __u16 = 11;
+pub const __NDA_MAX: __u16 = 12;
+
+impl Clone for ndmsg {
+    fn clone(&self) -> Self {
+        Self { ..*self }
+    }
+}
+
+impl Default for ndmsg {
+    fn default() -> Self {
+        unsafe { mem::zeroed::<Self>() }
+    }
+}
+
+#[repr(C)]
+#[derive(Copy)]
 pub struct nlmsgerr {
     pub error: libc::c_int,
     pub msg: nlmsghdr,
@@ -1397,122 +1433,6 @@ impl RtnlHandle {
         }
 
         Ok((slv, lv))
-    }
-
-    pub fn list_interfaces(&mut self) -> Result<Vec<Interface>> {
-        let mut ifaces: Vec<Interface> = Vec::new();
-
-        unsafe {
-            // get link info
-            let (_slv, lv) = self.dump_all_links()?;
-
-            // get addrinfo
-            let (_sav, av) = self.dump_all_addresses(0)?;
-
-            // got all the link message and address message
-            // into lv and av respectively, parse attributes
-            for link in &lv {
-                let nlh: *const nlmsghdr = *link;
-                let ifi: *const ifinfomsg = NLMSG_DATA!(nlh) as *const ifinfomsg;
-
-                if (*nlh).nlmsg_type != RTM_NEWLINK && (*nlh).nlmsg_type != RTM_DELLINK {
-                    continue;
-                }
-
-                if (*nlh).nlmsg_len < NLMSG_SPACE!(mem::size_of::<ifinfomsg>()) {
-                    info!(
-                        sl!(),
-                        "invalid nlmsg! nlmsg_len: {}, nlmsg_space: {}",
-                        (*nlh).nlmsg_len,
-                        NLMSG_SPACE!(mem::size_of::<ifinfomsg>())
-                    );
-                    break;
-                }
-
-                let rta: *mut rtattr = IFLA_RTA!(ifi) as *mut rtattr;
-                let rtalen = IFLA_PAYLOAD!(nlh) as u32;
-
-                let attrs = parse_attrs(rta, rtalen, (IFLA_MAX + 1) as usize)?;
-
-                // fill out some fields of Interface,
-                let mut iface: Interface = Interface::default();
-
-                if attrs[IFLA_IFNAME as usize] as i64 != 0 {
-                    let t = attrs[IFLA_IFNAME as usize];
-                    iface.name = String::from_utf8(getattr_var(t as *const rtattr))?;
-                }
-
-                if attrs[IFLA_MTU as usize] as i64 != 0 {
-                    let t = attrs[IFLA_MTU as usize];
-                    iface.mtu = getattr32(t) as u64;
-                }
-
-                if attrs[IFLA_ADDRESS as usize] as i64 != 0 {
-                    let alen = RTA_PAYLOAD!(attrs[IFLA_ADDRESS as usize]);
-                    let a: *const u8 = RTA_DATA!(attrs[IFLA_ADDRESS as usize]) as *const u8;
-                    iface.hwAddr = format_address(a, alen as u32)?;
-                }
-
-                // get ip address info from av
-                let mut ads: Vec<IPAddress> = Vec::new();
-
-                for address in &av {
-                    let alh: *const nlmsghdr = *address;
-                    let ifa: *const ifaddrmsg = NLMSG_DATA!(alh) as *const ifaddrmsg;
-                    let arta: *mut rtattr = IFA_RTA!(ifa) as *mut rtattr;
-
-                    if (*alh).nlmsg_type != RTM_NEWADDR {
-                        continue;
-                    }
-
-                    let tlen = NLMSG_SPACE!(mem::size_of::<ifaddrmsg>());
-                    if (*alh).nlmsg_len < tlen {
-                        info!(
-                            sl!(),
-                            "invalid nlmsg! nlmsg_len: {}, nlmsg_space: {}",
-                            (*alh).nlmsg_len,
-                            tlen
-                        );
-                        break;
-                    }
-
-                    let artalen = IFA_PAYLOAD!(alh) as u32;
-
-                    if (*ifa).ifa_index as u32 == (*ifi).ifi_index as u32 {
-                        // found target addresses
-                        // parse attributes and fill out Interface
-                        let addrs = parse_attrs(arta, artalen, (IFA_MAX + 1) as usize)?;
-                        // fill address field of Interface
-                        let mut one: IPAddress = IPAddress::default();
-                        let mut tattr: *const rtattr = addrs[IFA_LOCAL as usize];
-                        if addrs[IFA_ADDRESS as usize] as i64 != 0 {
-                            tattr = addrs[IFA_ADDRESS as usize];
-                        }
-
-                        one.mask = format!("{}", (*ifa).ifa_prefixlen);
-                        let a: *const u8 = RTA_DATA!(tattr) as *const u8;
-                        let alen = RTA_PAYLOAD!(tattr);
-                        one.family = IPFamily::v4;
-
-                        if (*ifa).ifa_family == libc::AF_INET6 as u8 {
-                            one.family = IPFamily::v6;
-                        }
-
-                        // only handle IPv4 for now
-                        // if (*ifa).ifa_family == libc::AF_INET as u8{
-                        one.address = format_address(a, alen as u32)?;
-                        //}
-
-                        ads.push(one);
-                    }
-                }
-
-                iface.IPAddresses = RepeatedField::from_vec(ads);
-                ifaces.push(iface);
-            }
-        }
-
-        Ok(ifaces)
     }
 
     unsafe fn dump_all_links(&mut self) -> Result<(Vec<Vec<u8>>, Vec<*const nlmsghdr>)> {
@@ -1873,107 +1793,6 @@ impl RtnlHandle {
         Ok(())
     }
 
-    pub fn update_interface(&mut self, iface: &Interface) -> Result<Interface> {
-        // the reliable way to find link is using hardware address
-        // as filter. However, hardware filter might not be supported
-        // by netlink, we may have to dump link list and the find the
-        // target link. filter using name or family is supported, but
-        // we cannot use that to find target link.
-        // let's try if hardware address filter works. -_-
-
-        let ifinfo = self.find_link_by_hwaddr(iface.hwAddr.as_str())?;
-
-        // bring down interface if it is up
-
-        if ifinfo.ifi_flags & libc::IFF_UP as u32 != 0 {
-            self.set_link_status(&ifinfo, false)?;
-        }
-
-        // delete all addresses associated with the link
-        let del_addrs: Vec<RtIPAddr> = self.get_link_addresses(&ifinfo)?;
-
-        self.delete_all_addrs(&ifinfo, del_addrs.as_ref())?;
-
-        // add new ip addresses in request
-        for grpc_addr in &iface.IPAddresses {
-            let rtip = RtIPAddr::from(grpc_addr.clone());
-            self.add_one_address(&ifinfo, &rtip)?;
-        }
-
-        // set name, set mtu, IFF_NOARP. in one rtnl_talk.
-        let mut v: Vec<u8> = vec![0; 2048];
-        unsafe {
-            let p: *mut u8 = v.as_mut_ptr() as *mut u8;
-            let mut nlh: *mut nlmsghdr = p as *mut nlmsghdr;
-            let mut ifi: *mut ifinfomsg = NLMSG_DATA!(nlh) as *mut ifinfomsg;
-
-            (*nlh).nlmsg_len = NLMSG_LENGTH!(mem::size_of::<ifinfomsg>() as u32) as __u32;
-            (*nlh).nlmsg_type = RTM_NEWLINK;
-            (*nlh).nlmsg_flags = NLM_F_REQUEST;
-
-            self.seq += 1;
-            (*nlh).nlmsg_seq = self.seq;
-
-            (*ifi).ifi_family = ifinfo.ifi_family;
-            (*ifi).ifi_type = ifinfo.ifi_type;
-            (*ifi).ifi_index = ifinfo.ifi_index;
-
-            if iface.raw_flags & libc::IFF_NOARP as u32 != 0 {
-                (*ifi).ifi_change |= libc::IFF_NOARP as u32;
-                (*ifi).ifi_flags |= libc::IFF_NOARP as u32;
-            }
-
-            addattr32(nlh, IFLA_MTU, iface.mtu as u32);
-
-            // if str is null terminated, use addattr_var
-            // otherwise, use addattr_str
-            addattr_var(
-                nlh,
-                IFLA_IFNAME,
-                iface.name.as_ptr() as *const u8,
-                iface.name.len(),
-            );
-            // addattr_str(nlh, IFLA_IFNAME, iface.name.as_str());
-        }
-
-        self.rtnl_talk(v.as_mut_slice(), false)?;
-
-        let _ = self.set_link_status(&ifinfo, true);
-        // test remove this link
-        // let _ = self.remove_interface(iface)?;
-
-        Ok(iface.clone())
-        //return Err(ErrorKind::Nix(nix::Error::Sys(
-        //	Errno::EOPNOTSUPP)).into());
-    }
-
-    fn remove_interface(&mut self, iface: &Interface) -> Result<Interface> {
-        let ifinfo = self.find_link_by_hwaddr(iface.hwAddr.as_str())?;
-        self.set_link_status(&ifinfo, false)?;
-
-        // delete this link per request
-        let mut v: Vec<u8> = vec![0; 2048];
-        unsafe {
-            let mut nlh: *mut nlmsghdr = v.as_mut_ptr() as *mut nlmsghdr;
-            let mut ifi: *mut ifinfomsg = NLMSG_DATA!(nlh) as *mut ifinfomsg;
-            // No attributes needed?
-            (*nlh).nlmsg_len = NLMSG_LENGTH!(mem::size_of::<ifinfomsg>()) as __u32;
-            (*nlh).nlmsg_type = RTM_DELLINK;
-            (*nlh).nlmsg_flags = NLM_F_REQUEST;
-
-            self.seq += 1;
-            (*nlh).nlmsg_seq = self.seq;
-
-            (*ifi).ifi_family = ifinfo.ifi_family;
-            (*ifi).ifi_index = ifinfo.ifi_index;
-            (*ifi).ifi_type = ifinfo.ifi_type;
-
-            self.rtnl_talk(v.as_mut_slice(), false)?;
-        }
-
-        Ok(iface.clone())
-    }
-
     fn get_name_by_index(&mut self, index: i32) -> Result<String> {
         let mut v: Vec<u8> = vec![0; 2048];
         let mut i = 0;
@@ -2029,132 +1848,6 @@ impl RtnlHandle {
         }
 
         nix_errno(Errno::ENOENT)
-    }
-
-    pub fn list_routes(&mut self) -> Result<Vec<Route>> {
-        // currently, only dump routes from main table for ipv4
-        // ie, rtmsg.rtmsg_family = AF_INET, set RT_TABLE_MAIN
-        // attribute in dump request
-        // Fix Me: think about othe tables, ipv6..
-        let mut rs: Vec<Route> = Vec::new();
-
-        unsafe {
-            let (_srv, rv) = self.dump_all_route_msgs()?;
-
-            // parse out routes and store in rs
-            for r in &rv {
-                let nlh: *const nlmsghdr = *r;
-                let rtm: *const rtmsg = NLMSG_DATA!(nlh) as *const rtmsg;
-
-                if (*nlh).nlmsg_type != RTM_NEWROUTE && (*nlh).nlmsg_type != RTM_DELROUTE {
-                    info!(sl!(), "not route message!");
-                    continue;
-                }
-
-                let tlen = NLMSG_SPACE!(mem::size_of::<rtmsg>());
-                if (*nlh).nlmsg_len < tlen {
-                    info!(
-                        sl!(),
-                        "invalid nlmsg! nlmsg_len: {}, nlmsg_spae: {}",
-                        (*nlh).nlmsg_len,
-                        tlen
-                    );
-                    break;
-                }
-
-                let rta: *mut rtattr = RTM_RTA!(rtm) as *mut rtattr;
-
-                if (*rtm).rtm_table != RT_TABLE_MAIN as u8 {
-                    continue;
-                }
-
-                let rtalen = RTM_PAYLOAD!(nlh) as u32;
-
-                let attrs = parse_attrs(rta, rtalen, (RTA_MAX + 1) as usize)?;
-
-                let t = attrs[RTA_TABLE as usize];
-                if t as i64 != 0 {
-                    let table = getattr32(t);
-                    if table != RT_TABLE_MAIN {
-                        continue;
-                    }
-                }
-                // find source, destination, gateway, scope, and
-                // and device name
-
-                let mut t = attrs[RTA_DST as usize];
-                let mut rte: Route = Route::default();
-
-                // destination
-                if t as i64 != 0 {
-                    let data: *const u8 = RTA_DATA!(t) as *const u8;
-                    let len = RTA_PAYLOAD!(t) as u32;
-                    rte.dest = format!("{}/{}", format_address(data, len)?, (*rtm).rtm_dst_len);
-                }
-
-                // gateway
-                t = attrs[RTA_GATEWAY as usize];
-                if t as i64 != 0 {
-                    let data: *const u8 = RTA_DATA!(t) as *const u8;
-                    let len = RTA_PAYLOAD!(t) as u32;
-                    rte.gateway = format_address(data, len)?;
-
-                    // for gateway, destination is 0.0.0.0
-                    rte.dest = "0.0.0.0".to_string();
-                }
-
-                // source
-                t = attrs[RTA_SRC as usize];
-
-                if t as i64 == 0 {
-                    t = attrs[RTA_PREFSRC as usize];
-                }
-
-                if t as i64 != 0 {
-                    let data: *const u8 = RTA_DATA!(t) as *const u8;
-                    let len = RTA_PAYLOAD!(t) as u32;
-
-                    rte.source = format_address(data, len)?;
-
-                    if (*rtm).rtm_src_len != 0 {
-                        rte.source = format!("{}/{}", rte.source.as_str(), (*rtm).rtm_src_len);
-                    }
-                }
-
-                // scope
-                rte.scope = (*rtm).rtm_scope as u32;
-
-                // oif
-                t = attrs[RTA_OIF as usize];
-                if t as i64 != 0 {
-                    let data: *const i32 = RTA_DATA!(t) as *const i32;
-                    assert_eq!(RTA_PAYLOAD!(t), 4);
-
-                    /*
-
-                    let mut n: Vec<u8> = vec![0; libc::IF_NAMESIZE];
-                    let np: *mut libc::c_char = n.as_mut_ptr() as *mut libc::c_char;
-                    let tn = libc::if_indextoname(*data as u32,
-                        np);
-
-                    if tn as i64 == 0 {
-                        info!(sl!(), "no name?");
-                    } else {
-                        info!(sl!(), "name(indextoname): {}", String::from_utf8(n)?);
-                    }
-                    // std::process::exit(-1);
-                    */
-
-                    rte.device = self
-                        .get_name_by_index(*data)
-                        .unwrap_or("unknown".to_string());
-                }
-
-                rs.push(rte);
-            }
-        }
-
-        Ok(rs)
     }
 
     unsafe fn dump_all_route_msgs(&mut self) -> Result<(Vec<Vec<u8>>, Vec<*const nlmsghdr>)> {
@@ -2430,33 +2123,6 @@ impl RtnlHandle {
         }
 
         Ok(())
-    }
-
-    pub fn update_routes(&mut self, rt: &Vec<Route>) -> Result<Vec<Route>> {
-        let rs = self.get_all_routes()?;
-        self.delete_all_routes(&rs)?;
-
-        for grpcroute in rt {
-            if grpcroute.gateway.as_str() == "" {
-                let r = RtRoute::from(grpcroute.clone());
-                if r.index == -1 {
-                    continue;
-                }
-                self.add_one_route(&r)?;
-            }
-        }
-
-        for grpcroute in rt {
-            if grpcroute.gateway.as_str() != "" {
-                let r = RtRoute::from(grpcroute.clone());
-                if r.index == -1 {
-                    continue;
-                }
-                self.add_one_route(&r)?;
-            }
-        }
-
-        Ok(rt.clone())
     }
 
     pub fn handle_localhost(&mut self) -> Result<()> {
@@ -2762,6 +2428,20 @@ fn parse_cider(s: &str) -> Result<(Vec<u8>, u8)> {
     Ok((parse_ipaddr(addr.as_str())?, mask))
 }
 
+fn parse_addr(ip_address: &str) -> Result<(__u8, Vec<u8>)> {
+    let ip_data = parse_ipaddr(ip_address)?;
+    let family: __u8;
+
+    // ipv6
+    if ip_data.len() == 16 {
+        family = libc::AF_INET6 as __u8;
+    } else {
+        family = libc::AF_INET as __u8;
+    }
+
+    Ok((family, ip_data))
+}
+
 fn parse_mac(hwaddr: &str) -> Result<Vec<u8>> {
     let mut hw: Vec<u8> = vec![0; 6];
 
@@ -2780,82 +2460,10 @@ fn parse_mac(hwaddr: &str) -> Result<Vec<u8>> {
     Ok(hw)
 }
 
-impl From<Route> for RtRoute {
-    fn from(r: Route) -> Self {
-        // only handle ipv4
-
-        let index = {
-            let mut rh = RtnlHandle::new(NETLINK_ROUTE, 0).unwrap();
-            match rh.find_link_by_name(r.device.as_str()) {
-                Ok(ifi) => ifi.ifi_index,
-                Err(_) => -1,
-            }
-        };
-
-        let (dest, dst_len) = if r.dest.is_empty() {
-            (Some(vec![0 as u8; 4]), 0)
-        } else {
-            let (dst, mask) = parse_cider(r.dest.as_str()).unwrap();
-            (Some(dst), mask)
-        };
-
-        let (source, src_len) = if r.source.is_empty() {
-            (None, 0)
-        } else {
-            let (src, mask) = parse_cider(r.source.as_str()).unwrap();
-            (Some(src), mask)
-        };
-
-        let gateway = if r.gateway.is_empty() {
-            None
-        } else {
-            Some(parse_ipaddr(r.gateway.as_str()).unwrap())
-        };
-
-        /*
-                let (dest, dst_len) = if gateway.is_some() {
-                    (vec![0 as u8; 4], 0)
-                } else {
-                    (tdest, tdst_len)
-                };
-        */
-        Self {
-            dest,
-            source,
-            src_len,
-            dst_len,
-            index,
-            gateway,
-            scope: r.scope as u8,
-            protocol: RTPROTO_UNSPEC,
-        }
-    }
-}
-
 pub struct RtIPAddr {
     pub ip_family: __u8,
     pub ip_mask: __u8,
     pub addr: Vec<u8>,
-}
-
-impl From<IPAddress> for RtIPAddr {
-    fn from(ipi: IPAddress) -> Self {
-        let ip_family = if ipi.family == IPFamily::v4 {
-            libc::AF_INET
-        } else {
-            libc::AF_INET6
-        } as __u8;
-
-        let ip_mask = scan_fmt!(ipi.mask.as_str(), "{}", u8).unwrap();
-
-        let addr = parse_ipaddr(ipi.address.as_ref()).unwrap();
-
-        Self {
-            ip_family,
-            ip_mask,
-            addr,
-        }
-    }
 }
 
 #[inline]
